warc & recorder refactor: split BaseWARCWriter from MultiWARCWriter, move to warc/warcwriter.py, recorder/multifilewarcwriter.py split indexing functionality from base warc iterator, move to archiveindexer.py
